refactor: registration pages
