+++ src/pages/RegistrationPages/SignInPage/SignInPage.tsx
@@ -2,24 +2,25 @@ import { Form, Formik } from 'formik'
 import { Link, useNavigate } from 'react-router-dom'
 import TextInput from 'shared/inputs/TextInputs/TextInput'
 import PasswordInput from 'shared/inputs/PasswordInput/PasswordInput'
-import { signInErrorType } from './signInEntity'
 import VisitorBackgroundContainer from 'shared/containers/VisitorBackgroundContainer/VisitorBackgroundContainer'
 import VisitorPageTitle from 'shared/typography/VisitorPageTitle'
 import ButtonTFMain from 'shared/buttons/ButtonTFMain/ButtonTFMain'
-import { signInSchema } from './signInSchema'
 import { useState } from 'react'
-import { checkSignInErrorStatus } from './checkSignInError'
-import { signIn } from './signInUtils'
+import { checkRegistrationError } from '../checkRegistrationError'
+import { signIn } from '../registrationUtils'
 import ButtonTFDisabled from 'shared/buttons/ButtonTFDisabled/ButtonTFDisabled'
+import { RegistrationErrorType } from '../registrationEntity'
+import { signInSchema } from '../registrationSchemas'
 
 type Props = {}
 
 const SignInPage = (props: Props) => {
   let navigate = useNavigate()
-  const [signInErrorData, setSignInErrorData] = useState<signInErrorType>()
+  const [registrationErrorData, setRegistrationErrorData] =
+    useState<RegistrationErrorType>()
 
   return (
-    <div className="mx-0 my-0 h-full  md:mx-24 md:my-40">
+    <div className="mx-0 my-0 h-full md:mx-24 md:my-40">
       <VisitorBackgroundContainer>
         <Formik
           initialValues={{
@@ -28,7 +29,7 @@ const SignInPage = (props: Props) => {
           }}
           validationSchema={signInSchema}
           onSubmit={(values) => {
-            signIn(values, setSignInErrorData, navigate)
+            signIn(values, setRegistrationErrorData, navigate)
           }}
         >
           {(formik) => {
@@ -62,8 +63,8 @@ const SignInPage = (props: Props) => {
                       <ButtonTFMain label="Підтвердити" />
                     )}
 
-                    {signInErrorData
-                      ? checkSignInErrorStatus(signInErrorData)
+                    {registrationErrorData
+                      ? checkRegistrationError(registrationErrorData)
                       : null}
                   </div>
