--- src/pages/SignUpPage/signUpEntity.ts
@@ -1,6 +1,0 @@
-export type UserSignUpType = {
-  name: string
-  email: string
-  password: string
-  confirmedPassword: string
-}
