+++ tailwind.config.js
@@ -67,6 +67,7 @@ module.exports = {
         552: '552px',
         600: '600px',
         1392: '1392px',
+        'calc-h': 'calc(100vh - 80px)',
       },
       boxShadow: {
         page_shadow:
